requirements.txt
Django==2.0.4
djangorestframework==3.8.2
django-cors-headers==2.1.0
chardet==3.0.4
pytz==2018.4
six==1.11.0
django-oauth-toolkit==0.12.0